Move updateIncomeEvidence logic into IncomeEvidenceService
