+++ crime-evidence/src/main/java/uk/gov/justice/laa/crime/evidence/builder/CreateEvidenceDTOBuilder.java
@@ -4,9 +4,9 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Component;
 import uk.gov.justice.laa.crime.common.model.evidence.ApiCreateIncomeEvidenceRequest;
-import uk.gov.justice.laa.crime.common.model.evidence.ApiUpdateIncomeEvidenceRequest;
 import uk.gov.justice.laa.crime.evidence.dto.CreateEvidenceDTO;
-import uk.gov.justice.laa.crime.evidence.dto.UpdateEvidenceDTO;
+
+import java.math.BigDecimal;
 
 @Component
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
@@ -16,9 +16,9 @@ public class CreateEvidenceDTOBuilder {
         return CreateEvidenceDTO.builder()
             .magCourtOutcome(request.getMagCourtOutcome())
             .applicantDetails(request.getApplicantDetails())
-            .applicantPensionAmount(request.getApplicantPensionAmount() != null ? request.getApplicantPensionAmount().doubleValue() : 0)
+            .applicantPensionAmount(request.getApplicantPensionAmount() != null ? request.getApplicantPensionAmount() : BigDecimal.ZERO)
             .partnerDetails(request.getPartnerDetails())
-            .partnerPensionAmount(request.getPartnerPensionAmount() != null ? request.getPartnerPensionAmount().doubleValue() : 0)
+            .partnerPensionAmount(request.getPartnerPensionAmount() != null ? request.getPartnerPensionAmount() : BigDecimal.ZERO)
             .build();
     }
 }
