+++ crime-evidence/src/test/java/uk/gov/justice/laa/crime/evidence/service/IncomeEvidenceServiceTest.java
@@ -1,27 +1,43 @@
 package uk.gov.justice.laa.crime.evidence.service;
 
-import static org.mockito.Mockito.when;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.justice.laa.crime.common.model.evidence.ApiApplicantDetails;
+import uk.gov.justice.laa.crime.common.model.evidence.ApiCreateIncomeEvidenceResponse;
+import uk.gov.justice.laa.crime.common.model.evidence.ApiIncomeEvidence;
+import uk.gov.justice.laa.crime.common.model.evidence.ApiIncomeEvidenceItems;
+import uk.gov.justice.laa.crime.common.model.evidence.ApiUpdateIncomeEvidenceResponse;
+import uk.gov.justice.laa.crime.enums.EmploymentStatus;
+import uk.gov.justice.laa.crime.enums.MagCourtOutcome;
+import uk.gov.justice.laa.crime.enums.evidence.IncomeEvidenceType;
+import uk.gov.justice.laa.crime.evidence.data.builder.TestModelDataBuilder;
+import uk.gov.justice.laa.crime.evidence.dto.CreateEvidenceDTO;
+import uk.gov.justice.laa.crime.evidence.dto.EvidenceReceivedResultDTO;
+import uk.gov.justice.laa.crime.evidence.dto.UpdateEvidenceDTO;
+import uk.gov.justice.laa.crime.evidence.repository.IncomeEvidenceRequiredItemRepository;
+import uk.gov.justice.laa.crime.evidence.repository.IncomeEvidenceRequiredRepository;
+import uk.gov.justice.laa.crime.evidence.staticdata.entity.IncomeEvidenceRequiredEntity;
+import uk.gov.justice.laa.crime.evidence.staticdata.enums.ApplicantType;
+import uk.gov.justice.laa.crime.evidence.staticdata.projection.IncomeEvidenceRequiredItemProjection;
+import uk.gov.justice.laa.crime.util.DateUtil;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.justice.laa.crime.common.model.evidence.ApiIncomeEvidence;
-import uk.gov.justice.laa.crime.enums.EmploymentStatus;
-import uk.gov.justice.laa.crime.enums.MagCourtOutcome;
-import uk.gov.justice.laa.crime.enums.evidence.IncomeEvidenceType;
-import uk.gov.justice.laa.crime.evidence.dto.EvidenceReceivedResultDTO;
-import uk.gov.justice.laa.crime.evidence.repository.IncomeEvidenceRequiredItemRepository;
-import uk.gov.justice.laa.crime.evidence.repository.IncomeEvidenceRequiredRepository;
-import uk.gov.justice.laa.crime.evidence.staticdata.entity.IncomeEvidenceRequiredEntity;
-import uk.gov.justice.laa.crime.evidence.staticdata.enums.ApplicantType;
-import uk.gov.justice.laa.crime.evidence.staticdata.projection.IncomeEvidenceRequiredItemProjection;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class IncomeEvidenceServiceTest {
@@ -32,6 +48,12 @@ class IncomeEvidenceServiceTest {
     @Mock
     private IncomeEvidenceRequiredItemRepository incomeEvidenceRequiredItemRepository;
 
+    @Mock
+    private IncomeEvidenceValidationService incomeEvidenceValidationService;
+
+    @InjectMocks
+    private IncomeEvidenceService incomeEvidenceService;
+
     @Test
     void givenNoRequiredEvidenceItemsExist_whenIsRequiredEvidenceOutstandingIsInvokedWithNoProvidedEvidenceItems_thenReturnFalse() {
         when(incomeEvidenceRequiredItemRepository.findByIncomeEvidenceRequiredId(1)).thenReturn(
@@ -41,11 +63,9 @@ class IncomeEvidenceServiceTest {
             )
         );
 
-        IncomeEvidenceService incomeEvidenceService = new IncomeEvidenceService(incomeEvidenceRequiredRepository, incomeEvidenceRequiredItemRepository);
-
         boolean result = incomeEvidenceService.isRequiredEvidenceOutstanding(1, Collections.emptyList());
 
-        Assertions.assertFalse(result);
+        assertFalse(result);
     }
 
     @Test
@@ -62,11 +82,9 @@ class IncomeEvidenceServiceTest {
             new ApiIncomeEvidence(35, null, IncomeEvidenceType.BANK_STATEMENT, false, "Bank statement")
         );
 
-        IncomeEvidenceService incomeEvidenceService = new IncomeEvidenceService(incomeEvidenceRequiredRepository, incomeEvidenceRequiredItemRepository);
-
         boolean result = incomeEvidenceService.isRequiredEvidenceOutstanding(1, evidenceItems);
 
-        Assertions.assertFalse(result);
+        assertFalse(result);
     }
 
     @Test
@@ -78,11 +96,9 @@ class IncomeEvidenceServiceTest {
             )
         );
 
-        IncomeEvidenceService incomeEvidenceService = new IncomeEvidenceService(incomeEvidenceRequiredRepository, incomeEvidenceRequiredItemRepository);
-
         boolean result = incomeEvidenceService.isRequiredEvidenceOutstanding(1, Collections.emptyList());
 
-        Assertions.assertTrue(result);
+        assertTrue(result);
     }
 
     @Test
@@ -101,11 +117,9 @@ class IncomeEvidenceServiceTest {
             new ApiIncomeEvidence(3, null, IncomeEvidenceType.EMP_LETTER, true, "Employment letter")
         );
 
-        IncomeEvidenceService incomeEvidenceService = new IncomeEvidenceService(incomeEvidenceRequiredRepository, incomeEvidenceRequiredItemRepository);
-
         boolean result = incomeEvidenceService.isRequiredEvidenceOutstanding(2, evidenceItems);
 
-        Assertions.assertTrue(result);
+        assertTrue(result);
     }
 
     @Test
@@ -113,8 +127,6 @@ class IncomeEvidenceServiceTest {
         when(incomeEvidenceRequiredRepository.getNumberOfEvidenceItemsRequired(MagCourtOutcome.APPEAL_TO_CC.getOutcome(), EmploymentStatus.EMPLOY.getCode(), EmploymentStatus.EMPLOYED_CASH.getCode(), "APPLICANT", 0d)).thenReturn(null);
 
         List<ApiIncomeEvidence> providedEvidenceItems = Collections.emptyList();
-
-        IncomeEvidenceService incomeEvidenceService = new IncomeEvidenceService(incomeEvidenceRequiredRepository, incomeEvidenceRequiredItemRepository);
 
         EvidenceReceivedResultDTO result = incomeEvidenceService.checkMinimumEvidenceItemsReceived(
             providedEvidenceItems,
@@ -124,8 +136,8 @@ class IncomeEvidenceServiceTest {
             EmploymentStatus.EMPLOYED_CASH,
             BigDecimal.ZERO);
 
-        Assertions.assertEquals(0, result.getMinimumEvidenceItemsRequired());
-        Assertions.assertTrue(result.isEvidenceReceived());
+        assertEquals(0, result.getMinimumEvidenceItemsRequired());
+        assertTrue(result.isEvidenceReceived());
     }
 
     @Test
@@ -136,8 +148,6 @@ class IncomeEvidenceServiceTest {
 
         List<ApiIncomeEvidence> providedEvidenceItems = Collections.emptyList();
 
-        IncomeEvidenceService incomeEvidenceService = new IncomeEvidenceService(incomeEvidenceRequiredRepository, incomeEvidenceRequiredItemRepository);
-
         EvidenceReceivedResultDTO result = incomeEvidenceService.checkMinimumEvidenceItemsReceived(
             providedEvidenceItems,
             ApplicantType.APPLICANT,
@@ -146,8 +156,8 @@ class IncomeEvidenceServiceTest {
             EmploymentStatus.EMPLOYED_CASH,
             BigDecimal.ZERO);
 
-        Assertions.assertEquals(0, result.getMinimumEvidenceItemsRequired());
-        Assertions.assertTrue(result.isEvidenceReceived());
+        assertEquals(0, result.getMinimumEvidenceItemsRequired());
+        assertTrue(result.isEvidenceReceived());
     }
 
     @Test
@@ -162,8 +172,6 @@ class IncomeEvidenceServiceTest {
             new ApiIncomeEvidence(3, LocalDate.of(2024, 9, 1), IncomeEvidenceType.EMP_LETTER, true, "Employment letter")
         );
 
-        IncomeEvidenceService incomeEvidenceService = new IncomeEvidenceService(incomeEvidenceRequiredRepository, incomeEvidenceRequiredItemRepository);
-
         EvidenceReceivedResultDTO result = incomeEvidenceService.checkMinimumEvidenceItemsReceived(
             providedEvidenceItems,
             ApplicantType.APPLICANT,
@@ -172,8 +180,8 @@ class IncomeEvidenceServiceTest {
             EmploymentStatus.EMPLOYED_CASH,
             BigDecimal.ZERO);
 
-        Assertions.assertEquals(2, result.getMinimumEvidenceItemsRequired());
-        Assertions.assertTrue(result.isEvidenceReceived());
+        assertEquals(2, result.getMinimumEvidenceItemsRequired());
+        assertTrue(result.isEvidenceReceived());
     }
 
     @Test
@@ -188,8 +196,6 @@ class IncomeEvidenceServiceTest {
             new ApiIncomeEvidence(3, LocalDate.of(2024, 9, 1), IncomeEvidenceType.EMP_LETTER, true, "Employment letter")
         );
 
-        IncomeEvidenceService incomeEvidenceService = new IncomeEvidenceService(incomeEvidenceRequiredRepository, incomeEvidenceRequiredItemRepository);
-
         EvidenceReceivedResultDTO result = incomeEvidenceService.checkMinimumEvidenceItemsReceived(
             providedEvidenceItems,
             ApplicantType.APPLICANT,
@@ -198,8 +204,8 @@ class IncomeEvidenceServiceTest {
             EmploymentStatus.EMPLOYED_CASH,
             BigDecimal.ZERO);
 
-        Assertions.assertEquals(4, result.getMinimumEvidenceItemsRequired());
-        Assertions.assertFalse(result.isEvidenceReceived());
+        assertEquals(4, result.getMinimumEvidenceItemsRequired());
+        assertFalse(result.isEvidenceReceived());
     }
 
 
@@ -211,8 +217,6 @@ class IncomeEvidenceServiceTest {
 
         List<ApiIncomeEvidence> applicantEvidenceItems = new ArrayList<>();
 
-        IncomeEvidenceService incomeEvidenceService = new IncomeEvidenceService(incomeEvidenceRequiredRepository, incomeEvidenceRequiredItemRepository);
-
         boolean result = incomeEvidenceService.checkEvidenceReceived(
             applicantEvidenceItems,
             MagCourtOutcome.APPEAL_TO_CC,
@@ -221,7 +225,7 @@ class IncomeEvidenceServiceTest {
             BigDecimal.ZERO,
             ApplicantType.APPLICANT);
 
-        Assertions.assertFalse(result);
+        assertFalse(result);
     }
 
     @Test
@@ -246,8 +250,6 @@ class IncomeEvidenceServiceTest {
             new ApiIncomeEvidence(3, null, IncomeEvidenceType.EMP_LETTER, true, "Employment letter")
         );
 
-        IncomeEvidenceService incomeEvidenceService = new IncomeEvidenceService(incomeEvidenceRequiredRepository, incomeEvidenceRequiredItemRepository);
-
         boolean result = incomeEvidenceService.checkEvidenceReceived(
             applicantEvidenceItems,
             MagCourtOutcome.APPEAL_TO_CC,
@@ -256,7 +258,7 @@ class IncomeEvidenceServiceTest {
             BigDecimal.ZERO,
             ApplicantType.APPLICANT);
 
-        Assertions.assertFalse(result);
+        assertFalse(result);
     }
 
     @Test
@@ -281,8 +283,6 @@ class IncomeEvidenceServiceTest {
             new ApiIncomeEvidence(36, LocalDate.of(2024, 9, 1), IncomeEvidenceType.EMP_LETTER, true, "Employment letter")
         );
 
-        IncomeEvidenceService incomeEvidenceService = new IncomeEvidenceService(incomeEvidenceRequiredRepository, incomeEvidenceRequiredItemRepository);
-
         boolean result = incomeEvidenceService.checkEvidenceReceived(
             applicantEvidenceItems,
             MagCourtOutcome.APPEAL_TO_CC,
@@ -291,7 +291,7 @@ class IncomeEvidenceServiceTest {
             BigDecimal.ZERO,
             ApplicantType.APPLICANT);
 
-        Assertions.assertTrue(result);
+        assertTrue(result);
     }
 
     private static IncomeEvidenceRequiredItemProjection createIncomeEvidenceRequiredItemProjection(
@@ -313,5 +313,214 @@ class IncomeEvidenceServiceTest {
                 return description;
             }
         };
+    }
+
+    @Test
+    void givenNoEvidenceItems_whenUpdateEvidenceIsInvoked_thenExceptionIsThrown() {
+        UpdateEvidenceDTO updateEvidenceDTO = TestModelDataBuilder.getUpdateEvidenceRequest();
+
+        assertThatThrownBy(() -> incomeEvidenceService.updateEvidence(updateEvidenceDTO))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("No income evidence items provided");
+    }
+
+    @Test
+    void givenValidationErrorOnEvidenceReceivedDate_whenUpdateEvidenceIsInvoked_thenExceptionIsThrown() {
+        LocalDate applicationReceivedDate = LocalDate.of(2024, 8, 30);
+        LocalDate evidenceItemReceivedDate = LocalDate.of(2024, 8, 29);
+        LocalDate existingEvidenceDueDate = LocalDate.of(2024, 9, 7);
+
+        List<ApiIncomeEvidence> applicantEvidenceItems = List.of(
+                new ApiIncomeEvidence(1, evidenceItemReceivedDate, IncomeEvidenceType.ACCOUNTS, false, "Company accounts")
+        );
+
+        UpdateEvidenceDTO updateEvidenceDTO =  TestModelDataBuilder.getUpdateEvidenceRequest(
+                applicationReceivedDate,
+                null,
+                applicantEvidenceItems,
+                false,
+                null,
+                evidenceItemReceivedDate,
+                existingEvidenceDueDate);
+
+        doThrow(IllegalArgumentException.class).when(incomeEvidenceValidationService).checkEvidenceReceivedDate(
+                DateUtil.parseLocalDate(updateEvidenceDTO.getEvidenceReceivedDate()), updateEvidenceDTO.getApplicationReceivedDate());
+
+        assertThatThrownBy(() -> incomeEvidenceService.updateEvidence(updateEvidenceDTO))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @Test
+    void givenValidationErrorOnEvidenceDueDate_whenUpdateEvidenceIsInvoked_thenExceptionIsThrown() {
+        LocalDate applicationReceivedDate = LocalDate.of(2024, 8, 30);
+        LocalDate evidenceItemReceivedDate = LocalDate.of(2024, 8, 31);
+        LocalDate existingEvidenceDueDate = LocalDate.of(2024, 9, 7);
+
+        List<ApiIncomeEvidence> applicantEvidenceItems = List.of(
+                new ApiIncomeEvidence(1, evidenceItemReceivedDate, IncomeEvidenceType.ACCOUNTS, false, "Company accounts")
+        );
+
+        UpdateEvidenceDTO updateEvidenceDTO = TestModelDataBuilder.getUpdateEvidenceRequest(
+                applicationReceivedDate,
+                null,
+                applicantEvidenceItems,
+                true,
+                null,
+                evidenceItemReceivedDate,
+                existingEvidenceDueDate);
+
+        doThrow(IllegalArgumentException.class).when(incomeEvidenceValidationService).checkEvidenceDueDates(
+                null,
+                existingEvidenceDueDate,
+                true);
+
+        assertThatThrownBy(() -> incomeEvidenceService.updateEvidence(updateEvidenceDTO))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @Test
+    void givenEvidenceHasNotBeenReceivedAndIncomeEvidenceHasReceivedDate_whenUpdateEvidenceIsInvoked_thenIncomeEvidenceReceivedDateIsRemoved() {
+        LocalDate applicationReceivedDate = LocalDate.of(2024, 8, 30);
+        LocalDate evidenceItemReceivedDate = LocalDate.of(2024, 9, 1);
+        LocalDate evidenceDueDate = LocalDate.of(2024, 9, 30);
+        LocalDate evidenceReceivedDate = LocalDate.of(2024, 9, 1);
+        LocalDate existingEvidenceDueDate = LocalDate.of(2024, 9, 7);
+
+        List<ApiIncomeEvidence> applicantEvidenceItems = List.of(
+                new ApiIncomeEvidence(1, evidenceItemReceivedDate, IncomeEvidenceType.ACCOUNTS, false, "Company accounts")
+        );
+
+        ApiApplicantDetails applicantDetails = buildApplicantDetails(1, EmploymentStatus.EMPLOY);
+
+        UpdateEvidenceDTO updateEvidenceDTO = TestModelDataBuilder.getUpdateEvidenceRequest(
+                applicationReceivedDate,
+                applicantDetails,
+                applicantEvidenceItems,
+                false,
+                evidenceDueDate,
+                evidenceReceivedDate,
+                existingEvidenceDueDate);
+
+        ApiUpdateIncomeEvidenceResponse expectedResponse = new ApiUpdateIncomeEvidenceResponse()
+                .withApplicantEvidenceItems(new ApiIncomeEvidenceItems(applicantDetails, applicantEvidenceItems))
+                .withDueDate(evidenceDueDate)
+                .withAllEvidenceReceivedDate(null);
+        when(incomeEvidenceRequiredRepository.getNumberOfEvidenceItemsRequired(any(), any(), any(), any(), any()))
+                .thenReturn(IncomeEvidenceRequiredEntity.builder().evidenceItemsRequired(3).build());
+
+        ApiUpdateIncomeEvidenceResponse actualResponse = incomeEvidenceService.updateEvidence(updateEvidenceDTO);
+
+        assertEquals(expectedResponse, actualResponse);
+    }
+
+    @Test
+    void givenOnlyPartnerEvidenceIsProvided_whenUpdateEvidenceIsInvoked_thenIncomeEvidenceIsUpdated() {
+        LocalDate applicationReceivedDate = LocalDate.of(2024, 8, 30);
+        LocalDate evidenceItemReceivedDate = LocalDate.of(2024, 9, 1);
+        LocalDate evidenceDueDate = LocalDate.of(2024, 9, 30);
+        LocalDate evidenceReceivedDate = LocalDate.of(2024, 9, 1);
+        LocalDate existingEvidenceDueDate = LocalDate.of(2024, 9, 7);
+
+        List<ApiIncomeEvidence> partnerEvidenceItems = List.of(
+                new ApiIncomeEvidence(1, evidenceItemReceivedDate, IncomeEvidenceType.ACCOUNTS, false, "Company accounts")
+        );
+
+        ApiApplicantDetails applicantDetails = buildApplicantDetails(1, EmploymentStatus.EMPLOY);
+        ApiApplicantDetails partnerDetails = buildApplicantDetails(2, EmploymentStatus.EMPLOYED_CASH);
+
+        UpdateEvidenceDTO updateEvidenceDTO = TestModelDataBuilder.getUpdateEvidenceRequest(
+                applicationReceivedDate,
+                applicantDetails,
+                null,
+                false,
+                evidenceDueDate,
+                evidenceReceivedDate,
+                existingEvidenceDueDate);
+        updateEvidenceDTO.setPartnerDetails(partnerDetails);
+        updateEvidenceDTO.setPartnerIncomeEvidenceItems(partnerEvidenceItems);
+
+        ApiUpdateIncomeEvidenceResponse expectedResponse = new ApiUpdateIncomeEvidenceResponse()
+                .withApplicantEvidenceItems(new ApiIncomeEvidenceItems(applicantDetails, Collections.emptyList()))
+                .withPartnerEvidenceItems(new ApiIncomeEvidenceItems(partnerDetails, partnerEvidenceItems))
+                .withDueDate(evidenceDueDate)
+                .withAllEvidenceReceivedDate(evidenceReceivedDate);
+
+        ApiUpdateIncomeEvidenceResponse actualResponse = incomeEvidenceService.updateEvidence(updateEvidenceDTO);
+
+        assertNotNull(updateEvidenceDTO.getEvidenceReceivedDate());
+        assertEquals(expectedResponse, actualResponse);
+    }
+
+    private ApiApplicantDetails buildApplicantDetails(int applicantId, EmploymentStatus employmentStatus) {
+        return new ApiApplicantDetails()
+                .withId(applicantId)
+                .withEmploymentStatus(employmentStatus);
+    }
+
+    @Test
+    void givenValidCreateEvidenceDTO_whenCreateEvidenceIsInvoked_thenDefaultIncomeEvidenceIsCreated() {
+        CreateEvidenceDTO createEvidenceDTO = TestModelDataBuilder.getCreateEvidenceRequest();
+        createEvidenceDTO.setPartnerDetails(null);
+        when(incomeEvidenceRequiredRepository.getNumberOfEvidenceItemsRequired(any(), any(), any(), any(), any()))
+                .thenReturn(IncomeEvidenceRequiredEntity.builder().evidenceItemsRequired(1).build());
+        when(incomeEvidenceRequiredItemRepository.findByIncomeEvidenceRequiredId(any()))
+                .thenReturn(List.of(createIncomeEvidenceRequiredItemProjection(1, IncomeEvidenceType.NINO.getName(), true)));
+        ApiIncomeEvidence apiIncomeEvidence = new ApiIncomeEvidence()
+                .withEvidenceType(IncomeEvidenceType.NINO)
+                .withMandatory(true);
+        ApiCreateIncomeEvidenceResponse expectedResponse = new ApiCreateIncomeEvidenceResponse()
+                .withApplicantEvidenceItems(getApiIncomeEvidenceItems(createEvidenceDTO.getApplicantDetails(), apiIncomeEvidence));
+
+        assertEquals(expectedResponse, incomeEvidenceService.createEvidence(createEvidenceDTO));
+    }
+
+    @Test
+    void givenValidCreateEvidenceDTOWithPartnerDetails_whenCreateEvidenceIsInvoked_thenDefaultIncomeEvidenceIsCreated() {
+        CreateEvidenceDTO createEvidenceDTO = TestModelDataBuilder.getCreateEvidenceRequest();
+        when(incomeEvidenceRequiredRepository.getNumberOfEvidenceItemsRequired(any(), any(), any(), any(), any()))
+                .thenReturn(IncomeEvidenceRequiredEntity.builder().evidenceItemsRequired(1).build());
+        when(incomeEvidenceRequiredItemRepository.findByIncomeEvidenceRequiredId(any()))
+                .thenReturn(List.of(createIncomeEvidenceRequiredItemProjection(1, IncomeEvidenceType.NINO.getName(), true)));
+        ApiIncomeEvidence apiIncomeEvidence = new ApiIncomeEvidence()
+                .withEvidenceType(IncomeEvidenceType.NINO)
+                .withMandatory(true);
+        ApiCreateIncomeEvidenceResponse expectedResponse = new ApiCreateIncomeEvidenceResponse()
+                .withPartnerEvidenceItems(getApiIncomeEvidenceItems(createEvidenceDTO.getPartnerDetails(), apiIncomeEvidence))
+                .withApplicantEvidenceItems(getApiIncomeEvidenceItems(createEvidenceDTO.getApplicantDetails(), apiIncomeEvidence));
+
+        assertEquals(expectedResponse, incomeEvidenceService.createEvidence(createEvidenceDTO));
+
+    }
+
+    private static ApiIncomeEvidenceItems getApiIncomeEvidenceItems(ApiApplicantDetails apiApplicantDetails, ApiIncomeEvidence apiIncomeEvidence) {
+        return new ApiIncomeEvidenceItems()
+                .withApplicantDetails(apiApplicantDetails)
+                .withIncomeEvidenceItems(List.of(apiIncomeEvidence));
+    }
+
+    @Test
+    void givenAScenarioWhereNullEvidenceItemsReturned_whenCreateEvidenceIsInvoked_thenNoDefaultIncomeEvidenceIsCreated() {
+        CreateEvidenceDTO createEvidenceDTO = TestModelDataBuilder.getCreateEvidenceRequest();
+        createEvidenceDTO.setPartnerDetails(null);
+        when(incomeEvidenceRequiredRepository.getNumberOfEvidenceItemsRequired(any(), any(), any(), any(), any()))
+                .thenReturn(null);
+        ApiCreateIncomeEvidenceResponse expectedResponse = new ApiCreateIncomeEvidenceResponse()
+                .withApplicantEvidenceItems(new ApiIncomeEvidenceItems()
+                        .withApplicantDetails(createEvidenceDTO.getApplicantDetails()));
+
+        assertEquals(expectedResponse, incomeEvidenceService.createEvidence(createEvidenceDTO));
+    }
+
+    @Test
+    void givenAScenarioWhereZeroEvidenceItemsReturned_whenCreateEvidenceIsInvoked_thenNoDefaultIncomeEvidenceIsCreated() {
+        CreateEvidenceDTO createEvidenceDTO = TestModelDataBuilder.getCreateEvidenceRequest();
+        createEvidenceDTO.setPartnerDetails(null);
+        when(incomeEvidenceRequiredRepository.getNumberOfEvidenceItemsRequired(any(), any(), any(), any(), any()))
+                .thenReturn(IncomeEvidenceRequiredEntity.builder().evidenceItemsRequired(0).build());
+        ApiCreateIncomeEvidenceResponse expectedResponse = new ApiCreateIncomeEvidenceResponse()
+                .withApplicantEvidenceItems(new ApiIncomeEvidenceItems()
+                        .withApplicantDetails(createEvidenceDTO.getApplicantDetails()));
+
+        assertEquals(expectedResponse, incomeEvidenceService.createEvidence(createEvidenceDTO));
     }
 }
